Update envfile.local.txt
I do not use Linux often and the Docker stuff is all new to me.  It took me embarrassing days to figure out that I needed to put the local IP for docker into the configuration.  Once I figured that out determining the IP was not achievable by the documented example of getting docker connection.